fix #17
I changed a `grep()` call with `perl = FALSE` (default) to `perl = TRUE`, is `perl` always available?

`rgnparser:::install_gnparser_bin()` has high chances to throw a warning but succeed in the next step of the loop so maybe warning should be shown only if all steps of the loops fail?

In `rgnparser:::bin_paths()` the call to `rgnparser:::pkg_file()` only adds an empty string to `path`. I don't see if this is useful.